# prop/portal/openapi/prop-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Prop.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/portal/signup/availability": {
      "get": {
        "tags": [
          "Signup"
        ],
        "parameters": [
          {
            "name": "firmId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal/signup": {
      "post": {
        "tags": [
          "Signup"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal/signup/verify": {
      "post": {
        "tags": [
          "Signup"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifySignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifySignupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal/login-help": {
      "post": {
        "tags": [
          "Signup"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvailabilityResponse": {
        "required": [
          "firmId",
          "available",
          "reason",
          "suggestions"
        ],
        "type": "object",
        "properties": {
          "firmId": {
            "type": "string"
          },
          "available": {
            "type": "boolean"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          },
          "suggestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "Whether the short name can be chosen, and when it is taken or reserved up to three free names like it."
      },
...
      "PasswordResetRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "The email of the person who forgot the password."
      },
...
      "SignupRequest": {
        "required": [
          "firmName",
          "firmId",
          "email",
          "password",
          "acceptTerms"
        ],
        "type": "object",
        "properties": {
          "firmName": {
            "type": [
              "null",
              "string"
            ]
          },
          "firmId": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": [
              "null",
              "string"
            ]
          },
          "acceptTerms": {
            "type": "boolean"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "FirmId is the short name: the portal's subdomain and the server on the trading platform.\nCurrency is the currency of the firm's accounts, one of the platform's, or its first when left out."
      },
      "SignupResponse": {
        "required": [
          "verificationRequired",
          "firmId",
          "adminUrl"
        ],
        "type": "object",
        "properties": {
          "verificationRequired": {
            "type": "boolean"
          },
          "firmId": {
            "type": "string"
          },
          "adminUrl": {
            "type": [
              "null",
              "string"
            ],
            "format": "uri"
          }
        },
        "description": "Either the email with the confirmation link is sent, or the firm is created and AdminUrl logs its administrator in."
      },
...
      "VerifySignupRequest": {
        "required": [
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "VerifySignupResponse": {
        "required": [
          "firmId",
          "adminUrl"
        ],
        "type": "object",
        "properties": {
          "firmId": {
            "type": "string"
          },
          "adminUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "description": "Open AdminUrl once to be logged in to the new firm's admin panel."
      },
...
    }
  },
...
}